+++ lib/src/async/ime.dart
@@ -12,21 +12,21 @@ class IME {
 
   Future<List<String>> getAvailableEngines() => _client.send(
       _handler.ime.buildGetAvailableEnginesRequest(),
-      _handler.ime.parseGetAvailableEngines);
+      _handler.ime.parseGetAvailableEnginesResponse);
 
   Future<String> getActiveEngine() => _client.send(
       _handler.ime.buildGetActiveEngineRequest(),
-      _handler.ime.parseGetActiveEngine);
+      _handler.ime.parseGetActiveEngineResponse);
 
   Future<bool> isActivated() => _client.send(
-      _handler.ime.buildIsActivatedRequest(), _handler.ime.parseIsActivated);
+      _handler.ime.buildIsActivatedRequest(), _handler.ime.parseIsActivatedResponse);
 
   Future<void> deactivate() => _client.send(
-      _handler.ime.buildDeactivateRequest(), _handler.ime.parseDeactivate);
+      _handler.ime.buildDeactivateRequest(), _handler.ime.parseDeactivateResponse);
 
   Future<void> activateEngine(String imeName) => _client.send(
       _handler.ime.buildActiveEngineRequest(imeName),
-      _handler.ime.parseActiveEngine);
+      _handler.ime.parseActiveEngineResponse);
 
   @override
   int get hashCode => _client.hashCode;