+++ lib/src/handler/w3c/keyboard.dart
@@ -1,5 +1,8 @@
 import 'package:appium_driver/src/common/webdriver_handler.dart';
 
+import 'package:appium_driver/src/common/w3c/command.dart';
+import 'package:appium_driver/src/common/request.dart';
+
 import 'package:webdriver/src/common/request.dart'; // ignore: implementation_imports
 import 'package:webdriver/src/handler/w3c/utils.dart'; // ignore: implementation_imports
 
@@ -48,4 +51,26 @@ class W3cKeyboardHandler extends KeyboardHandler {
   void parseSendKeysResponse(WebDriverResponse response) {
     parseW3cResponse(response);
   }
+
+  @override
+  WebDriverRequest buildHideKeyboardRequest() {
+    return AppiumWebDriverRequest.sendRequest(W3CCommands.HIDE_KEYBOARD);
+  }
+
+  @override
+  void parseHideKeyboardResponse(WebDriverResponse response) {
+    parseW3cResponse(response);
+  }
+
+  @override
+  WebDriverRequest buildIsKeyboardShownRequest() {
+    return AppiumWebDriverRequest.sendRequest(W3CCommands.IS_KEYBOARD_SHOWN);
+  }
+
+  @override
+  bool parseIsKeyboardShownResponse(WebDriverResponse response) {
+    return parseW3cResponse(response);
+  }
+
+
 }